First commit branch <dev>, implemented Maven project with consoleApp, with JUnit tests (with Mockito)
Refactored JUnit tests and providers

Refactored classes CharCounterResult, UniqueCharacter, MathProviderImpl, ViewProviderImpl, CharCounter, CharCounterConsoleApplication, MathProviderTest, ViewProviderTest

Added LRUCashe

Refactored and optimized

Added LRUCache

Optimized and refactored

Refactored Test classes (refactored old tests and add new tests) and other classes

Refactored and optimized

Refactored and Optimized

Optimized so that the CharCounterResult class is no longer needed and has been removed. A new implementation of LRUCache was also started (not finished). As a result of the changes, the tests are not relevant and cannot be passed (errors).

A new implementation of LRUCache was finished and fixed bugs charCounter. As a result of the changes, the tests are not relevant and cannot be passed (errors).

Refactoring and bug fixing. JUnit tests are described.

Refactored CharCounter

Refactored classes CharCounter and ViewProviderImpl and JUnit test class CharCounter.

Refactored class ViewProviderImpl